project structure was changed quest-storage-web module was added QuestStorageController was replaced by QuestController QuestBean was added
