+++ dao-manager/dao-manager-ejb/src/main/java/org/my/adventure/dao_manager/ejb/dao/QuestDAOImpl.java
@@ -15,6 +15,7 @@ import java.io.Serializable;
  */
 @Stateless
 @Remote(QuestDAO.class)
+//@Local
 public class QuestDAOImpl extends CommonDAOImpl<Quest> implements QuestDAO{
 
     public Quest getTestObject() {
