+++ quest-storage/pom.xml
@@ -18,6 +18,7 @@
         <module>quest-storage-api</module>
         <module>quest-storage-impl</module>
         <module>quest-storage-webapp</module>
+        <module>quest-storage-web</module>
     </modules>
 
 </project>
