+++ quest-storage/quest-storage-api/pom.xml
@@ -11,5 +11,4 @@
 
     <artifactId>quest-storage-api</artifactId>
 
-
 </project>
