+++ quest-storage/quest-storage-webapp/src/main/webapp/quest-storage/StorageMenu.xhtml
@@ -11,8 +11,8 @@
     <ui:composition template="/templates/menu.xhtml">
         <ui:define name="menu_right">
             <f:facet name="right">
-                <p:commandButton value="change layout"
-                                 actionListener="#{questStorageController.changeLayout()}" ajax="false"/>
+                <p:commandButton value="Изменить отображение"
+                                 actionListener="#{questController.changeLayout()}" ajax="false"/>
                 <p:commandButton type="button" title="Save" icon="ui-icon-disk"/>
                 <p:commandButton type="button" title="Delete" icon="ui-icon-trash"/>
                 <p:commandButton type="button" title="Print" icon="ui-icon-print"/>
