+++ quest-storage/quest-storage-webapp/src/main/webapp/quest-storage/StorageScroller.xhtml
@@ -11,9 +11,9 @@
 <h:body>
     <div align="center">
         <h:form>
-            <p:dataScroller value="#{questStorageController.quests}" var="quest" chunkSize="3" id="dataScroller">
+            <p:dataScroller value="#{questController.quests}" var="quest" chunkSize="3" id="dataScroller">
                 <f:facet name="header">
-                    Scroll Down to Load More Quests
+                    Прокрутите вниз, что бы увидеть больше квестов
                 </f:facet>
                 <div align="left">
                     <p:panelGrid style="width: 100%">
@@ -55,19 +55,19 @@
                         <f:facet name="footer">
                             <p:row>
                                 <p:column colspan="3">
-                                    <p:button outcome="/index.xhtml" value="Play" icon="ui-icon-star">
+                                    <p:button outcome="/index.xhtml" value="Играть" icon="ui-icon-star">
                                         <f:param name="questId" value="#{quest.id}" />
                                     </p:button>
-                                    <p:button outcome="/quest-editor/editor.xhtml" value="Edit" icon="ui-icon-star">
+                                    <p:button outcome="/quest-editor/editor.xhtml" value="Редактировать" icon="ui-icon-star">
                                         <f:param name="questId" value="#{quest.id}" />
                                     </p:button>
-                                    <p:commandButton value="Delete" actionListener="#{questStorageController.deleteQuest(quest.id)}" update="dataScroller">
-                                        <p:confirm header="Confirmation" message="Are you sure?" icon="ui-icon-alert" />
+                                    <p:commandButton value="Удалить" actionListener="#{questController.deleteQuest(quest.id)}" update="dataScroller">
+                                        <p:confirm header="Требуется подтверждение" message="Вы уверены?" icon="ui-icon-alert" />
                                     </p:commandButton>
 
                                     <p:confirmDialog global="true" showEffect="fade" hideEffect="fade" closeOnEscape="true" closable="true">
-                                        <p:commandButton value="No" type="button" styleClass="ui-confirmdialog-no" icon="ui-icon-close" />
-                                        <p:commandButton value="Yes" type="button" styleClass="ui-confirmdialog-yes" icon="ui-icon-check" />
+                                        <p:commandButton value="Нет" type="button" styleClass="ui-confirmdialog-no" icon="ui-icon-close" />
+                                        <p:commandButton value="Да" type="button" styleClass="ui-confirmdialog-yes" icon="ui-icon-check" />
                                     </p:confirmDialog>
                                 </p:column>
                             </p:row>
