+++ text-adventure-app/text-adventure-app-web/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,4 +1,4 @@
 <jboss-web>
-    <!--<security-domain>text-adventure-domain</security-domain>-->
-    <security-domain>java:/jaas/other</security-domain>
+    <security-domain>text-adventure-domain</security-domain>
+    <!--<security-domain>java:/jaas/other</security-domain>-->
 </jboss-web>
