+++ text-adventure-app/text-adventure-app-web/src/main/webapp/WEB-INF/web.xml
@@ -40,43 +40,22 @@
         <param-value>excite-bike</param-value>
     </context-param>
 
+    <!--Defining security constraint for type of roles available-->
+
     <security-constraint>
         <web-resource-collection>
             <web-resource-name>secured</web-resource-name>
             <url-pattern>/faces/quest-editor/*</url-pattern>
             <url-pattern>/quest-editor/*</url-pattern>
-            <http-method>GET</http-method>
             <http-method>POST</http-method>
+            <http-method>GET</http-method>
+            <http-method>PUT</http-method>
+            <http-method>DELETE</http-method>
         </web-resource-collection>
         <auth-constraint>
-            <role-name>MANAGER</role-name>
+            <role-name>admin</role-name>
         </auth-constraint>
     </security-constraint>
-
-    <security-role>
-        <role-name>MANAGER</role-name>
-    </security-role>
-
-    <login-config>
-        <auth-method>BASIC</auth-method>
-    </login-config>
-
-    <!--Defining security constraint for type of roles available-->
-
-    <!--<security-constraint>-->
-        <!--<web-resource-collection>-->
-            <!--<web-resource-name>secured</web-resource-name>-->
-            <!--<url-pattern>/faces/quest-editor/*</url-pattern>-->
-            <!--<url-pattern>/quest-editor/*</url-pattern>-->
-            <!--<http-method>POST</http-method>-->
-            <!--<http-method>GET</http-method>-->
-            <!--<http-method>PUT</http-method>-->
-            <!--<http-method>DELETE</http-method>-->
-        <!--</web-resource-collection>-->
-        <!--<auth-constraint>-->
-            <!--<role-name>ADMIN</role-name>-->
-        <!--</auth-constraint>-->
-    <!--</security-constraint>-->
 
     <!--<security-constraint>-->
         <!--<web-resource-collection>-->
@@ -95,24 +74,50 @@
     <!--Defining security constraint for type of roles available-->
 
     <!--Defining type of authenitcation mechanism-->
-    <!--<login-config>-->
-        <!--<auth-method>BASIC</auth-method>-->
-        <!--<realm-name>text-adventure-domain</realm-name>-->
-        <!--<form-login-config>-->
-            <!--<form-login-page>/authorization/login.xhtml</form-login-page>-->
-            <!--<form-error-page>/authorization/login_error.xhtml</form-error-page>-->
-        <!--</form-login-config>-->
-    <!--</login-config>-->
+    <login-config>
+        <auth-method>FORM</auth-method>
+        <realm-name>text-adventure-domain</realm-name>
+        <form-login-config>
+            <form-login-page>/authorization/login.xhtml</form-login-page>
+            <form-error-page>/authorization/login_error.xhtml</form-error-page>
+        </form-login-config>
+    </login-config>
     <!--Defining type of authenitcation mechanism-->
 
     <!--Denining security role-->
-    <!--<security-role>-->
-        <!--<role-name>ADMIN</role-name>-->
-    <!--</security-role>-->
+    <security-role>
+        <role-name>admin</role-name>
+    </security-role>
 
-    <!--<security-role>-->
-        <!--<role-name>user</role-name>-->
-    <!--</security-role>-->
+    <security-role>
+        <role-name>user</role-name>
+    </security-role>
     <!--Denining security role-->
 
+
+    <error-page>
+        <!-- Missing login -->
+        <error-code>401</error-code>
+        <location>/authorization/login_error.xhtml</location>
+    </error-page>
+    <error-page>
+        <!-- Forbidden directory listing -->
+        <error-code>403</error-code>
+        <location>/authorization/login_error.xhtml</location>
+    </error-page>
+    <!--<error-page>-->
+        <!--&lt;!&ndash; Uncaught exception &ndash;&gt;-->
+        <!--<error-code>500</error-code>-->
+        <!--<location>/authorization/login_error.xhtml</location>-->
+    <!--</error-page>-->
+    <!--<error-page>-->
+        <!--&lt;!&ndash; Unsupported servlet method &ndash;&gt;-->
+        <!--<error-code>503</error-code>-->
+        <!--<location>/authorization/login_error.xhtml</location>-->
+    <!--</error-page>-->
+    <error-page>
+        <!-- Not found -->
+        <error-code>404</error-code>
+        <location>/templates/error.xhtml</location>
+    </error-page>
 </web-app>
