+++ text-adventure-app/text-adventure-app-web/src/main/webapp/authorization/login.xhtml
@@ -6,7 +6,7 @@
       xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
       xmlns:f="http://xmlns.jcp.org/jsf/core">
 <h:body>
-    <form action="j_security_check" method="post">
+    <form action="j_security_check" method="post" style="text-align: center">
         Input for username:
         <input name="j_username" type="text"/>
         Input for password:
