+++ text-adventure-app/text-adventure-app-web/src/main/webapp/authorization/login_error.xhtml
@@ -4,10 +4,15 @@
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://xmlns.jcp.org/jsf/html"
       xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:p="http://primefaces.org/ui"
       xmlns:f="http://xmlns.jcp.org/jsf/core">
 <h:body>
     <h:form>
-        <h1>login failure!</h1>
+        <h1 align="center">Неверное имя пользователя или пароль, попробуйте войти ещё раз.</h1>
+        <h1 align="center">Wrong username or password, try again</h1>
+        <div align="center">
+            <p:button outcome="/index.xhtml" value="Quest Cracker" icon="ui-icon-star"/>
+        </div>
     </h:form>
 </h:body>
 </html>
