+++ text-adventure-app/text-adventure-app-web/src/main/webapp/templates/error.xhtml
@@ -4,8 +4,11 @@
 <html xmlns="http://www.w3.org/1999/xhtml"
       xmlns:h="http://xmlns.jcp.org/jsf/html"
       xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:p="http://primefaces.org/ui"
       xmlns:f="http://xmlns.jcp.org/jsf/core">
-<f:view>
-    <h:outputLabel value="Error!"/>
-</f:view>
+<h:body>
+
+    <h:outputLabel value="Упс, что-то пошло не так!" style="text-align: center" />
+    <p:button outcome="/index.xhtml" value="Quest Cracker" icon="ui-icon-star"/>
+</h:body>
 </html>
