+++ text-adventure-app/text-adventure-app-web/src/main/webapp/templates/menu.xhtml
@@ -11,18 +11,10 @@
         <p:toolbar id="tb">
             <ui:insert name="menu_left">
                 <f:facet name="left">
-                    <p:button outcome="/index.xhtml" value="Home" icon="ui-icon-star">
-                        <f:param name="questId" value="10" />
-                    </p:button>
-                    <p:button outcome="/quest-editor/editor.xhtml" value="Quest editor" icon="ui-icon-star">
-                        <f:param name="questId" value="10" />
-                    </p:button>
-                    <p:button outcome="/quest-storage/storage.xhtml" value="Quest storage" icon="ui-icon-star">
-                        <f:param name="questId" value="10" />
-                    </p:button>
-                    <p:button outcome="/authorization/login.xhtml" value="authorization" icon="ui-icon-star">
-                        <f:param name="questId" value="10" />
-                    </p:button>
+                    <p:button outcome="/index.xhtml" value="Quest Cracker" icon="ui-icon-star"/>
+                    <p:button outcome="/quest-editor/editor.xhtml" value="Редактор квестов" icon="ui-icon-star"/>
+                    <p:button outcome="/quest-storage/storage.xhtml" value="Список квестов" icon="ui-icon-star"/>
+                    <!--<p:button outcome="/authorization/login.xhtml" value="Войти" icon="ui-icon-star"/>-->
                     <!--<p:commandButton type="button" value="New" icon="ui-icon-document" />-->
                     <!--<p:commandButton type="button" value="Open" icon="ui-icon-folder-open"/>-->
                 </f:facet>
